clipboard-apis/copy_operation_does_not_dispatch_other_events_copy_on_inputtext.html

<!DOCTYPE html>
<html><head>
    <meta charset="UTF-8">
    <title>Clipboard API - copy operation does not dispatch other events   - copy on 'input_text'</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="support/_lib.js"></script>
    <script>
/** copy operation does not dispatch other events   */
/* Events: copy */
/* Test HTML: <form><input id="input_text" autofocus onfocus="this.select()" value="copied text"></form>*/
/* Targets: document 'input_text' */
function clipboard_api_test(e, test_obj){
  var otherEventsFired=[];
  var events=[ 'input', 'textInput', 'change' ];
  for( var i=0,ev; ev=events[i];i++ )e.currentTarget.addEventListener(ev, function(e){ otherEventsFired.push(e.type); }, false);
    test_obj.step_timeout(function(){
      result(function(){
        assert_equals(otherEventsFired.length, 0);
      });
    }, 50);
  e.target.blur(); // might provoke a change event, but should not
}

    </script>
</head>
<body>

    <p>FAILED (This TC requires JavaScript enabled)</p>
    <form><input id="input_text" autofocus onfocus="this.select()" value="copied text"></form>
    <div id="log"></div>
    <script>
    (function(){
        var t = async_test()
        var dataToPaste='';
        var extPassCond=null;
        var eventTarget='input_text';
        var eventType='copy';
        setup(function(){
                setupTest(t, eventTarget, eventType, dataToPaste, extPassCond );
            });
    })();
    </script>
</body></html>
